Complete rewrite v2.0
There is no easy way to handle this. It needed a complete rewrite. First
it used an unnecessary temp file and was unable to handle comments in
the front of the ini file. It also was not compatible with Mac OS X.

This is a complete departure from the original project. Mainly because
the authors felt that cross compatibility was an [exercise for the
consumer][1]. I felt this was easy enough fix that Mac OS X styles can
coexist with Posix and GNU styles.

[1]: https://github.com/chilladx/config-parser/issues/4#issuecomment-139536928